+++ addon/locales/en/components/fd-modal-input-message-box.js
@@ -1,0 +1,5 @@
+export default {
+  'placeholder': 'Input value',
+  'delete-stage-text': 'Enter stage name for confirm delete',
+  'delete-stage-not-confirmed': 'Delete stage not confirmed'
+};