+++ addon/locales/ru/components/fd-modal-input-message-box.js
@@ -1,0 +1,5 @@
+export default {
+  'placeholder': 'Введите значение',
+  'delete-stage-text': 'Для подтверждения удаления введите имя стадии',
+  'delete-stage-not-confirmed': 'Удаление стадии не подтверждено'
+};