+++ app/components/fd-modal-input-message-box.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-designer/components/fd-modal-input-message-box';